issue #8: build code added
