--- download_pcam.py
@@ -1,4 +1,0 @@
-import gdown
-
-if __name__ == "__main__":
-    gdown.download_folder(id="1gHou49cA1s5vua2V5L98Lt8TiWA3FrKB")
